Pz first

+++ sustainable-api/src/main/java/com/sustainable/sustainableapi/controllers/UserController.java
@@ -1,12 +1,16 @@
 package com.sustainable.sustainableapi.controllers;
 
+import com.sustainable.sustainableapi.model.dtos.SignInDto;
+import com.sustainable.sustainableapi.model.dtos.TokenDto;
 import com.sustainable.sustainableapi.model.dtos.UserDto;
 import com.sustainable.sustainableapi.model.entities.User;
 import com.sustainable.sustainableapi.services.UserService;
-import org.springframework.http.HttpStatus;
+import com.sustainable.sustainableapi.utils.AuthorizationUtil;
 import org.springframework.http.MediaType;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 import java.util.UUID;
 
@@ -20,6 +24,17 @@ public class UserController {
         this.userService = userService;
     }
 
+    @PostMapping("/signin")
+    public TokenDto login(@RequestBody @Valid SignInDto signInDto) {
+        return userService.signin(signInDto);
+    }
+
+    @PostMapping("/signup")
+    public TokenDto signup(@RequestBody @Valid UserDto user) {
+        return userService.signup(user);
+    }
+
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping
     public List<UserDto> getUsers(){
         return userService.getUsers();
@@ -27,22 +42,19 @@ public class UserController {
 
     @GetMapping("/{userId}")
     public UserDto getUser(@PathVariable UUID userId) {
+        AuthorizationUtil.checkUserIsHimselfOrAdmin(userId);
         return userService.getUser(userId);
-    }
-
-    @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
-    @ResponseStatus(HttpStatus.CREATED)
-    public UserDto createUser(@RequestBody UserDto user) {
-        return userService.createUser(user);
     }
 
     @PutMapping(value = "/{userId}", consumes = MediaType.APPLICATION_JSON_VALUE)
     public UserDto updateUser(@RequestBody User user,@PathVariable UUID userId){
+        AuthorizationUtil.checkUserIsHimselfOrAdmin(userId);
         return userService.updateUser(user, userId);
     }
 
     @DeleteMapping("/{userId}")
     public void deleteUser(@PathVariable UUID userId) {
+        AuthorizationUtil.checkUserIsHimselfOrAdmin(userId);
         userService.deleteUser(userId);
     }
 }
